fix delta kernels
