+++ pynestml/utils/ast_utils.py
@@ -19,7 +19,6 @@
 # You should have received a copy of the GNU General Public License
 # along with NEST.  If not, see <http://www.gnu.org/licenses/>.
 
-
 from typing import Any, Dict, Iterable, List, Mapping, Optional, Sequence, Set, Union
 
 import re
@@ -27,7 +26,6 @@ import sympy
 
 from pynestml.codegeneration.printers.ast_printer import ASTPrinter
 from pynestml.codegeneration.printers.cpp_variable_printer import CppVariablePrinter
-from pynestml.codegeneration.printers.nestml_simple_expression_printer_units_as_factors import NESTMLSimpleExpressionPrinterUnitsAsFactors
 from pynestml.frontend.frontend_configuration import FrontendConfiguration
 from pynestml.generated.PyNestMLLexer import PyNestMLLexer
 from pynestml.meta_model.ast_assignment import ASTAssignment
@@ -2191,104 +2189,6 @@ class ASTUtils:
                 equations_block.get_declarations().remove(decl)
 
     @classmethod
-    def get_delta_factors_from_convolutions(cls, model: ASTModel) -> dict:
-        r"""
-        For every occurrence of a convolution of the form `x^(n) = a * convolve(kernel, inport) + ...` where `kernel` is a delta function, add the element `(x^(n), inport) --> a` to the set.
-        """
-        print("get_delta_factors_from_convolutions")
-        delta_factors = {}
-
-        for equations_block in model.get_equations_blocks():
-            for ode_eq in equations_block.get_ode_equations():
-                var = ode_eq.get_lhs()
-                expr = ode_eq.get_rhs()
-                conv_calls = ASTUtils.get_convolve_function_calls(expr)
-                for conv_call in conv_calls:
-                    assert len(conv_call.args) == 2, "convolve() function call should have precisely two arguments: kernel and spike input port"
-                    kernel = conv_call.args[0]
-                    if ASTUtils.is_delta_kernel(model.get_kernel_by_name(kernel.get_variable().get_name())):
-                        inport = conv_call.args[1].get_variable()
-                        factor_str = ASTUtils.get_factor_str_from_expr_and_inport(expr, str(conv_call))
-                        assert factor_str
-                        delta_factors[(var, inport)] = factor_str
-
-        return delta_factors
-
-    @classmethod
-    def get_delta_factors_from_input_port_references(cls, model: ASTModel) -> dict:
-        r"""
-        For every occurrence of a convolution of the form ``x^(n) = a * inport + ...``, add the element `(x^(n), inport) --> a` to the set.
-        """
-        delta_factors = {}
-
-        spike_inports = model.get_spike_input_ports()
-        for equations_block in model.get_equations_blocks():
-            for ode_eq in equations_block.get_ode_equations():
-                var = ode_eq.get_lhs()
-                expr = ode_eq.get_rhs()
-
-                for inport_sym in spike_inports:
-                    inport_ = ASTUtils.get_input_port_by_name(model.get_input_blocks(), inport_sym.name)
-
-                    inport_var = ASTNodeFactory.create_ast_variable(inport_sym.name)
-                    inport_var.update_scope(equations_block.get_scope())
-
-                    factor_str = ASTUtils.get_factor_str_from_expr_and_inport(expr, inport_var.name)
-
-                    if factor_str:
-                        delta_factors[(var, inport_var)] = factor_str
-
-                    # XXX: what about vectors?????
-
-        return delta_factors
-
-    @classmethod
-    def get_factor_str_from_expr_and_inport(cls, expr, sub_expr):
-        r"""Use ``sympy.coeff()`` to extract the factor ``a`` with which ``sub_expr`` appears in ``expr = ... + a * sub_expr + ...``"""
-
-        from sympy import sympify
-
-        from pynestml.codegeneration.printers.constant_printer import ConstantPrinter
-        from pynestml.codegeneration.printers.nestml_function_call_printer import NESTMLFunctionCallPrinter
-        from pynestml.codegeneration.printers.nestml_printer import NESTMLPrinter
-        from pynestml.codegeneration.printers.ode_toolbox_expression_printer import ODEToolboxExpressionPrinter
-        from pynestml.codegeneration.printers.ode_toolbox_variable_printer import ODEToolboxVariablePrinter
-
-        printer = NESTMLPrinter()
-        printer._expression_printer = ODEToolboxExpressionPrinter(simple_expression_printer=None)
-        printer._constant_printer = ConstantPrinter()
-        printer._function_call_printer = NESTMLFunctionCallPrinter(expression_printer=printer._expression_printer)
-        printer._variable_printer = ODEToolboxVariablePrinter(expression_printer=printer._expression_printer)
-        printer._simple_expression_printer = NESTMLSimpleExpressionPrinterUnitsAsFactors(variable_printer=printer._variable_printer, function_call_printer=printer._function_call_printer, constant_printer=printer._constant_printer)
-        printer._expression_printer._simple_expression_printer = printer._simple_expression_printer
-
-        expr_str = printer.print(expr)
-
-        root_node = expr
-        while not isinstance(root_node, ASTModel):
-            root_node = root_node.get_parent()
-
-        all_variable_symbols = root_node.get_parameter_symbols() + root_node.get_state_symbols() + root_node.get_internal_symbols()
-        all_variable_symbols_dict = {s.name: sympy.Symbol(s.name) for s in all_variable_symbols}
-
-        # pretend that physical units are variables
-        unit_vars_str = PredefinedUnits.get_units().keys()
-        all_variable_symbols_dict |= {s: sympy.Symbol(s) for s in unit_vars_str}
-
-        sympy_expr = sympify(expr_str, locals=all_variable_symbols_dict)  # minimal dict to make no assumptions (e.g. "beta" could otherwise be recognised as a function instead of as a parameter symbol)
-        sympy_expr = sympy.expand(sympy_expr)
-        sympy_sub_expr = sympy.parsing.sympy_parser.parse_expr(sub_expr)
-        factor_str = []
-        for term in sympy.Add.make_args(sympy_expr):
-            coeff = term.coeff(sympy_sub_expr)
-            if coeff:
-                factor_str.append(str(coeff))
-
-        factor_str = " + ".join(factor_str)
-
-        return factor_str
-
-    @classmethod
     def remove_kernel_definitions_from_equations_block(cls, model: ASTModel) -> Set[ASTDeclaration]:
         r"""
         Removes all kernels in equations blocks.
@@ -2305,44 +2205,6 @@ class ASTUtils:
                 equations_block.get_declarations().remove(decl)
 
         return all_removed_decls
-
-    @classmethod
-    def generate_kernel_buffers(cls, model: ASTModel, equations_block: Union[ASTEquationsBlock, List[ASTEquationsBlock]]) -> Mapping[ASTKernel, ASTInputPort]:
-        """
-        For every occurrence of a convolution of the form `convolve(var, spike_buf)`: add the element `(kernel, spike_buf)` to the set, with `kernel` being the kernel that contains variable `var`.
-        """
-
-        kernel_buffers = set()
-        convolve_calls = ASTUtils.get_convolve_function_calls(equations_block)
-        for convolve in convolve_calls:
-            el = (convolve.get_args()[0], convolve.get_args()[1])
-            sym = convolve.get_args()[0].get_scope().resolve_to_symbol(convolve.get_args()[0].get_variable().name, SymbolKind.VARIABLE)
-            if sym is None:
-                raise Exception("No initial value(s) defined for kernel with variable \""
-                                + convolve.get_args()[0].get_variable().get_complete_name() + "\"")
-            if sym.block_type == BlockType.INPUT:
-                # swap the order
-                el = (el[1], el[0])
-
-            # find the corresponding kernel object
-            var = el[0].get_variable()
-            assert var is not None
-            kernel = model.get_kernel_by_name(var.get_name())
-            assert kernel is not None, "In convolution \"convolve(" + str(var.name) + ", " + str(el[1]) + ")\": no kernel by name \"" + var.get_name() + "\" found in model."
-
-            el = (kernel, el[1])
-
-            # make sure no duplicates are added -- expressions should be compared as strings
-            el_exists = False
-            for _el in kernel_buffers:
-                if el[0] == _el[0] and str(el[1]) == str(_el[1]):
-                    el_exists = True
-                    break
-
-            if not el_exists:
-                kernel_buffers.add(el)
-
-        return kernel_buffers
 
     @classmethod
     def replace_convolution_aliasing_inlines(cls, neuron: ASTModel) -> None:
@@ -2405,56 +2267,6 @@ class ASTUtils:
             return replace_var(x)
 
         model.accept(ASTHigherOrderVisitor(func))
-
-    @classmethod
-    def replace_convolve_calls_with_buffers_(cls, model: ASTModel, equations_block: ASTEquationsBlock) -> None:
-        r"""
-        Replace all occurrences of `convolve(kernel[']^n, spike_input_port)` with the corresponding buffer variable, e.g. `g_E__X__spikes_exc[__d]^n` for a kernel named `g_E` and a spike input port named `spikes_exc`.
-        """
-        from pynestml.visitors.ast_symbol_table_visitor import ASTSymbolTableVisitor
-
-        def replace_function_call_through_var(_expr=None):
-            if _expr.is_function_call() and _expr.get_function_call().get_name() == "convolve":
-                convolve = _expr.get_function_call()
-                el = (convolve.get_args()[0], convolve.get_args()[1])
-                sym = convolve.get_args()[0].get_scope().resolve_to_symbol(
-                    convolve.get_args()[0].get_variable().name, SymbolKind.VARIABLE)
-                if sym.block_type == BlockType.INPUT:
-                    # swap elements
-                    el = (el[1], el[0])
-                var = el[0].get_variable()
-                spike_input_port = el[1].get_variable()
-                kernel = model.get_kernel_by_name(var.get_name())
-
-                _expr.set_function_call(None)
-                buffer_var = cls.construct_kernel_X_spike_buf_name(
-                    var.get_name(), spike_input_port, var.get_differential_order() - 1)
-                if cls.is_delta_kernel(kernel):
-                    # delta kernels are treated separately, and should be kept out of the dynamics (computing derivates etc.) --> set to zero
-                    _expr.set_variable(None)
-                    _expr.set_numeric_literal(0)
-                else:
-                    ast_variable = ASTVariable(buffer_var)
-                    ast_variable.set_source_position(_expr.get_source_position())
-                    _expr.set_variable(ast_variable)
-
-        def func_replace_function_call_through_var(x):
-            return replace_function_call_through_var(x) if isinstance(x, ASTSimpleExpression) else True
-
-        def replace_spiking_input_port_with_zero(_expr=None):
-            if _expr.get_variable():
-                port = ASTUtils.get_input_port_by_name(model.get_input_blocks(), _expr.get_variable().name)
-                if port and port.is_spike():
-                    # the spiking input port appears directly in the ODE -- this is equivalent to convolving with a delta kernel. Updates due to this term will be handled in ``delta_factors``
-                    _expr.set_variable(None)
-                    _expr.set_numeric_literal(0)
-
-        def func_replace_spiking_input_port_with_zero(x):
-            return replace_spiking_input_port_with_zero(x) if isinstance(x, ASTSimpleExpression) else True
-
-        equations_block.accept(ASTHigherOrderVisitor(func_replace_function_call_through_var))
-        equations_block.accept(ASTHigherOrderVisitor(func_replace_spiking_input_port_with_zero))
-        equations_block.accept(ASTSymbolTableVisitor())
 
     @classmethod
     def update_blocktype_for_common_parameters(cls, node):
@@ -2712,7 +2524,7 @@ class ASTUtils:
         return None
 
     @classmethod
-    def collect_variables_affected_by_ports(cls, model, post_port_names, strictly_synaptic_vars: Optional[Set[str]] = None):
+    def collect_variables_affected_by_ports(cls, model, post_port_names, metadata: Dict[str, Dict[str, Any]], strictly_synaptic_vars: Optional[Set[str]] = None):
 
         if not strictly_synaptic_vars:
             strictly_synaptic_vars: Set[str] = set()
@@ -2735,10 +2547,7 @@ class ASTUtils:
         # add names of convolutions
         all_state_vars += ASTUtils.get_all_variables_used_in_convolutions(model.get_equations_blocks(), model)
 
-        # add names of kernels
-        kernel_buffers = ASTUtils.generate_kernel_buffers(model, model.get_equations_blocks())
-        all_state_vars += [var.name for k in kernel_buffers for var in k[0].variables]
-
+        # exclude variables that are not assigned to in postsynaptic spike handler onReceive blocks
         for input_block in model.get_input_blocks():
             for port in input_block.get_input_ports():
                 if port.name not in post_port_names:
